requirements.txt
flask
flask_sqlalchemy
flask_wtf
wtforms
flask-login
Flask-Migrate
pandas
openpyxl
xlsxwriter
requests
python-dateutil
pypdf
reportlab
waitress
pytest
